refactor(tui): split event_dialog.go below 1000 lines
Move keys, update, and view into sibling files in package tui.
No behavior change.

Related: #652
